기능수정: 내 플레이리스트에 노래 추가 및 삭제 기능

+++ src/main/java/project/singasong/playlist/controller/PlaylistController.java
@@ -27,14 +27,14 @@ public class PlaylistController {
     private final PlaylistService playlistService;
 
     @GetMapping("/playlist")
-    public String playlist(Model model, HttpServletRequest request, @RequestParam(defaultValue = "0") Long offset) {
+    public String playlist(Model model, HttpServletRequest request, @RequestParam(defaultValue = "0") long offset) {
         Long userId = (Long) request.getSession().getAttribute("userId");
 
         if(userId == null) {
             return "redirect:/";
         }
 
-        List<Playlist> playlist = playlistService.getFindByUserId(PlaylistPagingDto.of(userId, offset));
+        List<Playlist> playlist = playlistService.findByUserId(PlaylistPagingDto.of(userId, offset));
         model.addAttribute("playlist", playlist);
         model.addAttribute("offset", playlist.get(playlist.size()-1).getId());
 
@@ -47,14 +47,19 @@ public class PlaylistController {
     }
 
     @GetMapping("/playlist-update/{playlistId}")
-    public String playlistUpdate(@PathVariable Long playlistId, Model model) {
+    public String playlistUpdate(@PathVariable Long playlistId, HttpServletRequest request,Model model) {
+        Long userId = (Long) request.getSession().getAttribute("userId");
+        if(!userId.equals(playlistService.findById(playlistId).getUserId())) {
+            return "playlist";
+        }
+
         model.addAttribute("playlistId", playlistId);
         return "playlist-update";
     }
 
     @GetMapping("/playlist/{userId}")
-    public ResponseEntity getFindByUserId(Model model, @PathVariable Long userId, @RequestParam(defaultValue = "0") Long offset) {
-        List<Playlist> playlist = playlistService.getFindByUserId(PlaylistPagingDto.of(userId, offset));
+    public ResponseEntity findByUserId(Model model, @PathVariable Long userId, @RequestParam(defaultValue = "0") long offset) {
+        List<Playlist> playlist = playlistService.findByUserId(PlaylistPagingDto.of(userId, offset));
 
         if(!playlist.isEmpty()) {
             model.addAttribute("playlist", playlist);
@@ -75,7 +80,12 @@ public class PlaylistController {
     }
 
     @PatchMapping("/playlist/{id}")
-    public @ResponseBody ResponseEntity update(@PathVariable Long id, @RequestBody UpdatePlaylistDto updateSongList) {
+    public @ResponseBody ResponseEntity update(@PathVariable Long id, @RequestBody UpdatePlaylistDto updateSongList, HttpServletRequest request) {
+        Long userId = (Long) request.getSession().getAttribute("userId");
+        if(!userId.equals(playlistService.findById(id).getUserId())) {
+            return ResponseEntity.ok().build();
+        }
+
         Playlist playlist = Playlist.builder()
             .id(id)
             .title(updateSongList.getTitle())
@@ -85,7 +95,12 @@ public class PlaylistController {
     }
 
     @DeleteMapping("/playlist/{id}")
-    public @ResponseBody ResponseEntity delete(@PathVariable Long id) {
+    public @ResponseBody ResponseEntity delete(@PathVariable Long id, HttpServletRequest request) {
+        Long userId = (Long) request.getSession().getAttribute("userId");
+        if(!userId.equals(playlistService.findById(id).getUserId())) {
+            return ResponseEntity.ok().build();
+        }
+
         return ResponseEntity.ok().body(playlistService.delete(id));
     }
